some more of the refactor
